Call listLikes function from index.js
