+++ src/listItems.js
@@ -1,6 +1,7 @@
 import fetchData from './fetchData.js';
+import { getAPI } from './involvementAPI.js';
 
-const listItems = async () => {
+export const listItems = async () => {
   const ul = document.querySelector('.recipes');
   const food = await fetchData();
   food.forEach((element) => {
@@ -39,4 +40,18 @@ const listItems = async () => {
   });
 };
 
-export default listItems;
+export const listLikes = () => {
+  getAPI().then((response) => {
+    const like = Array.from(document.querySelectorAll('.likes'));
+    for (let i = 0; i < like.length; i += 1) {
+      response.forEach((item) => {
+        if (item.item_id === like[i].id) {
+          const small = document.createElement('small');
+          small.className = ('small');
+          small.textContent = `${item.likes} likes`;
+          like[i].appendChild(small);
+        }
+      });
+    }
+  });
+};
